kinect_fetch_fw: use readlink -m when setting FW_DESTDIR
This way the normalization succeeds even when the parent directory in
"$1" does not exists.

For example, if FW_DESTDIR is going to be set to /lib/firmware/kinect
but /lib/firmware does not exist yet, there is no reason to fail as the
install command below is going to create all the needed sub-paths.